+++ web-frontend/modules/database/mixins/copyPasteHelper.js
@@ -2,11 +2,19 @@
  * A mixin that can be used to copy and paste row values.
  */
 
-import { setRichClipboard } from '@baserow/modules/database/utils/clipboard'
+import {
+  getRichClipboard,
+  setRichClipboard,
+  LOCAL_STORAGE_CLIPBOARD_KEY,
+} from '@baserow/modules/database/utils/clipboard'
+
+const PAPA_CONFIG = {
+  delimiter: '\t',
+}
 
 export default {
   methods: {
-    copySelectionToClipboard(fields, rows) {
+    prepareSelectionForCopy(fields, rows) {
       const textData = []
       const jsonData = []
       for (const row of rows) {
@@ -23,42 +31,65 @@ export default {
         textData.push(text)
         jsonData.push(json)
       }
-      const tsv = this.$papa.unparse(textData, {
-        delimiter: '\t',
-      })
-      const values = {
-        'text/plain': tsv,
-        'application/json': JSON.stringify(jsonData),
+      const text = this.$papa.unparse(textData, PAPA_CONFIG)
+      try {
+        localStorage.setItem(
+          LOCAL_STORAGE_CLIPBOARD_KEY,
+          JSON.stringify({ text, json: jsonData })
+        )
+        return text
+      } catch (e) {
+        // If the local storage is full then we just ignore it.
+        // @TODO: Should we warn the user?
+        return ''
       }
-      setRichClipboard(values)
+    },
+    async copySelectionToClipboard(selectionPromise) {
+      // Firefox does not have the ClipboardItem type enabled by default, so we
+      // need to check if it is available. Safari instead, needs the
+      // ClipboardItem type to save async data to the clipboard.
+      if (typeof ClipboardItem !== 'undefined') {
+        navigator.clipboard.write([
+          // eslint-disable-next-line no-undef
+          new ClipboardItem({
+            'text/plain': selectionPromise.then(
+              ([fields, rows]) =>
+                new Blob([this.prepareSelectionForCopy(fields, rows)], {
+                  type: 'text/plain',
+                })
+            ),
+          }),
+        ])
+      } else {
+        const text = await selectionPromise.then(([fields, rows]) =>
+          this.prepareSelectionForCopy(fields, rows)
+        )
+        if (typeof navigator.clipboard?.writeText !== 'undefined') {
+          navigator.clipboard.writeText(text)
+        } else {
+          setRichClipboard({ 'text/plain': text })
+        }
+      }
     },
     async extractClipboardData(event) {
-      const textRawData = event.clipboardData.getData('text/plain').trim()
-
-      let jsonRawData
-      if (event.clipboardData.types.includes('application/json')) {
-        jsonRawData = event.clipboardData.getData('application/json')
-      }
-
-      const { data: textData } = await this.$papa.parsePromise(textRawData, {
-        delimiter: '\t',
-      })
+      const { textRawData, jsonRawData } = await getRichClipboard(event)
+      const { data: textData } = await this.$papa.parsePromise(
+        textRawData,
+        PAPA_CONFIG
+      )
 
       let jsonData = null
-      if (jsonRawData) {
-        try {
-          const parsed = JSON.parse(jsonRawData)
-          // Check if we have an array of arrays with At least one row with at least
-          // one row with a value Otherwise the paste is empty
-          if (
-            Array.isArray(parsed) &&
-            parsed.length === textData.length &&
-            parsed.every((row) => Array.isArray(row)) &&
-            parsed.some((row, index) => row.length > 0)
-          ) {
-            jsonData = JSON.parse(jsonRawData)
-          }
-        } catch (e) {}
+      if (jsonRawData != null) {
+        // Check if we have an array of arrays with At least one row with at least
+        // one row with a value Otherwise the paste is empty
+        if (
+          Array.isArray(jsonRawData) &&
+          jsonRawData.length === textData.length &&
+          jsonRawData.every((row) => Array.isArray(row)) &&
+          jsonRawData.some((row) => row.length > 0)
+        ) {
+          jsonData = jsonRawData
+        }
       }
 
       return [textData, jsonData]